requirements.txt
# pip install -r requirements.txt
numpy>=1.21.0
glfw>=2.5.0
PyOpenGL>=3.1.6
pyopencl>=2021.1.0
pybind11>=2.6.0
